Document kind of path from FANTOIR doc
The FANTOIR documentation contains a table with all
the "nature de la voie" information.

This change provides the data and a mechanism to exploit it.

To update nature_voie.dat, just copy/paste from the table from the
documentation PDF works. Then, `utils/parse-kv-spaces` will parse
it to JSON.

Provide a Makefile to take care of those steps.